--- a/packages/babel-parser/src/plugins/jsx/index.js
+++ b/packages/babel-parser/src/plugins/jsx/index.js
@@ -574,6 +574,11 @@ export default (superClass: Class<Parser>): Class<Parser> =>
       }
     }
 
+    skipSpace() {
+      const curContext = this.curContext();
+      if (!curContext.preserveSpace) super.skipSpace();
+    }
+
     getTokenFromCode(code: number): void {
       const context = this.curContext();