… benchmark/babel-parser/many-nested-block-elements/bench.mjs
… benchmark/babel-parser/many-nested-jsx-elements-with-attributes/bench.mjs
… benchmark/babel-parser/many-nested-template-elements/bench.mjs
… benchmark/babel-parser/many-template-elements/bench.mjs
… packages/babel-parser/src/plugins/typescript/index.js
… packages/babel-parser/src/tokenizer/context.js
… packages/babel-parser/src/tokenizer/index.js
… packages/babel-parser/src/tokenizer/types.js
… packages/babel-parser/src/util/location.js
… packages/babel-parser/test/fixtures/es2015/template/trailing-comments/input.js
… packages/babel-parser/test/fixtures/es2015/template/trailing-comments/output.json
… packages/babel-parser/test/fixtures/tokens/template-string-babel-7/basic/input.js
… packages/babel-parser/test/fixtures/tokens/template-string-babel-7/basic/output.json
… packages/babel-parser/test/fixtures/tokens/template-string-babel-7/options.json
… packages/babel-parser/test/fixtures/tokens/template-string/basic/input.js
… packages/babel-parser/test/fixtures/tokens/template-string/basic/output.json
… packages/babel-parser/test/fixtures/tokens/template-string/options.json