<!--
Before making a PR, please read our contributing guidelines
https://github.com/babel/babel/blob/main/CONTRIBUTING.md

Please note that the Babel Team requires two approvals before merging most PRs.

For issue references: Add a comma-separated list of a [closing word](https://help.github.com/articles/closing-issues-via-commit-messages/) followed by the ticket number fixed by the PR. (it should be underlined in the preview if done correctly)

If you are making a change that should have a docs update: submit another PR to https://github.com/babel/website
-->

| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Major: Breaking Change?  | Yes, behind `BABEL_8_BREAKING` flag
| Tests Added + Pass?      | Yes
| License                  | MIT

<!-- Describe your changes below in as much detail as possible -->
This PR includes commits from https://github.com/babel/babel/pull/13891.

Let's start by an example. Currently Babel tokenizes a string template
```js
`before${x}middle${y}after`
```
into a token sequences
```
[ backQuote, template, dollarBraceL, name, braceR, template, dollarBraceL, name, braceR, template, backQuote ]
```

in order to check whether a backQuote <code>&#x60;</code> starts a template and whether braceR `}` starts a [template span](https://tc39.es/ecma262/#prod-TemplateSpans), we have to introduce a [context stack](https://github.com/babel/babel/blob/76e4566bfca8b299c28d8c96c30feb750f129016/packages/babel-parser/src/tokenizer/index.js#L1630-L1667) for backQuote and braceR. Unlike backQuote, a braceR can pair with more than one tokens: namely braceL `{`, dollarBraceL `${` and braceHashL `#{`. Since we lost track of what the braceR is paring with when tokenzing, we have to push new context for braceL and braceHashL even if they have no special semantics in string template. Given that braceL `{` is a common token in JavaScript, this approach introduced overhead that could have been avoided.

In this PR we merge braceR and dollarBraceL with the template token. Two new template token types are introduced in order to differentiate between template middle `...${` (ends with `${`) and template tail <code>...&#x60;</code> (ends with <code>&#x60;</code>). So now the example above is tokenized into

```
[ templateMiddle, name, templateMiddle, name, templateTail ]
```

The new token layout is terser, requires no context tracking and can still work with the parser. So we can remove `updateContext` in the base parser and also simplify the JSX element which still depends on `updateContext`.

Another bonus is that the new token layout is almost identical to the `espree` one. So we can also remove some works in `@babel/eslint-parser`: In Babel 8 we don't need to merge multiple tokens into a template token.

## Benchmark Results
<details>
<summary>many-template-elements (15% faster)</summary>

```
baseline 128 template elements: 8_172 ops/sec ±44.79% (0.122ms)
baseline 256 template elements: 5_256 ops/sec ±0.77% (0.19ms)
baseline 512 template elements: 2_660 ops/sec ±0.73% (0.376ms)
baseline 1024 template elements: 1_278 ops/sec ±1.59% (0.783ms)
current 128 template elements: 9_080 ops/sec ±44.47% (0.11ms)
current 256 template elements: 6_025 ops/sec ±0.85% (0.166ms)
current 512 template elements: 3_053 ops/sec ±0.27% (0.328ms)
current 1024 template elements: 1_442 ops/sec ±1.5% (0.693ms)
```
</details>

<details>
<summary>many-nested-template-elements (25% faster) </summary>

```
baseline 128 nested template elements: 7_857 ops/sec ±41.2% (0.127ms)
baseline 256 nested template elements: 4_737 ops/sec ±2.01% (0.211ms)
baseline 512 nested template elements: 2_247 ops/sec ±1.84% (0.445ms)
baseline 1024 nested template elements: 1_016 ops/sec ±1.87% (0.985ms)
current 128 nested template elements: 10_305 ops/sec ±31.59% (0.097ms)
current 256 nested template elements: 5_999 ops/sec ±1.81% (0.167ms)
current 512 nested template elements: 2_919 ops/sec ±0.98% (0.343ms)
current 1024 nested template elements: 1_331 ops/sec ±2.32% (0.751ms)
```
</details>

<a href="https://gitpod.io/#https://github.com/babel/babel/pull/13919"><img src="https://gitpod.io/button/open-in-gitpod.svg"/></a>